refactor: consolidate action routing to ActionRegistry
